requirements.txt
requests>=2.32.3
fpdf2>=2.7.9
